+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-# 2020-05-22 1.10.71
+# 2020-05-22 1.10.71-72
   - Auto-generates "All Tracks Playlist"
   - Hotspot password
+  - Calculate Track time remaining
 # 2020-03-11 1.10.67-70
   - Clear passcode on GPIO hold
   - Prevent crash if trying to generate thumbnail and thr file is (somehow) missing